Rename actionpack-page_caching to actionpack-page_caching_multithread
