+++ actionpack-page_caching_multithread.gemspec
@@ -1,13 +1,13 @@
 # -*- encoding: utf-8 -*-
 
 Gem::Specification.new do |gem|
-  gem.name          = 'actionpack-page_caching'
+  gem.name          = 'actionpack-page_caching_multithread'
   gem.version       = '1.0.2'
   gem.author        = 'David Heinemeier Hansson'
   gem.email         = 'david@loudthinking.com'
   gem.description   = 'Static page caching for Action Pack (removed from core in Rails 4.0)'
   gem.summary       = 'Static page caching for Action Pack (removed from core in Rails 4.0)'
-  gem.homepage      = 'https://github.com/rails/actionpack-page_caching'
+  gem.homepage      = 'https://github.com/dimelo/actionpack-page_caching_multithread'
   gem.license       = 'MIT'
 
   gem.files         = `git ls-files`.split($/)
