+++ lib/action_controller/page_caching_multithread.rb
