--- lib/actionpack/page_caching.rb
@@ -1,1 +1,0 @@
-require 'actionpack/page_caching/railtie'
