+++ lib/actionpack/page_caching_multithread.rb
@@ -1,0 +1,1 @@
+require 'actionpack/page_caching_multithread/railtie'
